Drop legacy version, refactor pages
