--- apps/website/src/components/settings/CourseDetails.stories.tsx
@@ -1,163 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react';
-import { createMockCourse, createMockCourseRegistration } from '../../__tests__/testUtils';
-import type { GroupDiscussionWithGroupAndUnit } from '../../server/routers/group-discussions';
-import CourseDetails from './CourseDetails';
-import { ONE_HOUR_SECONDS, ONE_MINUTE_SECONDS } from '../../lib/constants';
-
-const courseId = 'course-1';
-const mockCourse = createMockCourse({ id: courseId });
-const mockCourseRegistration = createMockCourseRegistration({ courseId });
-
-const now = Math.floor(Date.now() / 1000);
-
-const mockDiscussions: Record<string, GroupDiscussionWithGroupAndUnit> = {
-  'discussion-1': {
-    id: 'discussion-1',
-    facilitators: ['facilitator-1'],
-    participantsExpected: ['participant-1'],
-    attendees: [],
-    startDateTime: now + 2 * ONE_HOUR_SECONDS,
-    endDateTime: now + 3 * ONE_HOUR_SECONDS,
-    group: 'group-1',
-    zoomAccount: null,
-    courseSite: null,
-    unitNumber: 1,
-    unit: null,
-    unitFallback: '1: Introduction to AI Safety',
-    zoomLink: 'https://zoom.us/j/123',
-    activityDoc: 'https://docs.google.com/document/d/discussion-1-doc',
-    slackChannelId: null,
-    round: null,
-    courseBuilderUnitRecordId: 'unit-1',
-    autoNumberId: null,
-    unitRecord: { unitNumber: '1', title: 'Introduction to AI Safety' } as GroupDiscussionWithGroupAndUnit['unitRecord'],
-    groupDetails: { groupName: 'Group A' } as GroupDiscussionWithGroupAndUnit['groupDetails'],
-  },
-  'discussion-2': {
-    id: 'discussion-2',
-    facilitators: ['facilitator-1'],
-    participantsExpected: ['participant-1'],
-    attendees: [],
-    startDateTime: now + 7 * 24 * ONE_HOUR_SECONDS,
-    endDateTime: now + 7 * 24 * ONE_HOUR_SECONDS + ONE_HOUR_SECONDS,
-    group: 'group-1',
-    zoomAccount: null,
-    courseSite: null,
-    unitNumber: 2,
-    unit: null,
-    unitFallback: '2: AI Alignment',
-    zoomLink: 'https://zoom.us/j/456',
-    activityDoc: 'https://docs.google.com/document/d/discussion-2-doc',
-    slackChannelId: null,
-    round: null,
-    courseBuilderUnitRecordId: 'unit-2',
-    autoNumberId: null,
-    unitRecord: { unitNumber: '2', title: 'AI Alignment' } as GroupDiscussionWithGroupAndUnit['unitRecord'],
-    groupDetails: { groupName: 'Group A' } as GroupDiscussionWithGroupAndUnit['groupDetails'],
-  },
-  'discussion-3': {
-    id: 'discussion-3',
-    facilitators: ['facilitator-1'],
-    participantsExpected: ['participant-1'],
-    attendees: ['participant-1'],
-    startDateTime: now - 7 * 24 * ONE_HOUR_SECONDS,
-    endDateTime: now - 7 * 24 * ONE_HOUR_SECONDS + ONE_HOUR_SECONDS,
-    group: 'group-1',
-    zoomAccount: null,
-    courseSite: null,
-    unitNumber: 0,
-    unit: null,
-    unitFallback: '0: Kickoff',
-    zoomLink: 'https://zoom.us/j/789',
-    activityDoc: 'https://docs.google.com/document/d/discussion-3-doc',
-    slackChannelId: null,
-    round: null,
-    courseBuilderUnitRecordId: 'unit-0',
-    autoNumberId: null,
-    unitRecord: { unitNumber: '0', title: 'Kickoff' } as GroupDiscussionWithGroupAndUnit['unitRecord'],
-    groupDetails: { groupName: 'Group A' } as GroupDiscussionWithGroupAndUnit['groupDetails'],
-  },
-};
-
-const meta: Meta<typeof CourseDetails> = {
-  title: 'Settings/CourseDetails',
-  component: CourseDetails,
-  parameters: {
-    layout: 'padded',
-  },
-  args: {
-    course: mockCourse,
-    attendedDiscussions: [mockDiscussions['discussion-3']!],
-    upcomingDiscussions: [mockDiscussions['discussion-1']!, mockDiscussions['discussion-2']!],
-    facilitatedDiscussions: [],
-    isLoading: false,
-  },
-};
-
-export default meta;
-type Story = StoryObj<typeof meta>;
-
-export const Default: Story = {
-  args: {
-    courseRegistration: { ...mockCourseRegistration, role: 'Participant' },
-  },
-};
-
-export const ManyUpcomingDiscussions: Story = {
-  args: {
-    courseRegistration: { ...mockCourseRegistration, role: 'Participant' },
-    upcomingDiscussions: Array.from({ length: 5 }, (_, index) => {
-      const unitNumber = index + 1;
-      return {
-        ...mockDiscussions['discussion-1']!,
-        id: `discussion-${unitNumber}`,
-        startDateTime: now + (unitNumber + 1) * 7 * 24 * ONE_HOUR_SECONDS,
-        endDateTime: now + (unitNumber + 1) * 7 * 24 * ONE_HOUR_SECONDS + ONE_HOUR_SECONDS,
-        unitNumber,
-        unitFallback: `${unitNumber}: Unit ${unitNumber}`,
-        unitRecord: {
-          unitNumber: unitNumber.toString(),
-          title: `Unit ${unitNumber}`,
-        } as GroupDiscussionWithGroupAndUnit['unitRecord'],
-      };
-    }),
-  },
-};
-
-export const Facilitator: Story = {
-  args: {
-    courseRegistration: { ...mockCourseRegistration, role: 'Facilitator' },
-    attendedDiscussions: [],
-    facilitatedDiscussions: [mockDiscussions['discussion-3']!],
-  },
-  parameters: {
-    docs: {
-      description: {
-        story: 'Facilitators see "Facilitated discussions" tab instead of "Attended discussions".',
-      },
-    },
-  },
-};
-
-export const LiveDiscussion: Story = {
-  args: {
-    courseRegistration: { ...mockCourseRegistration, role: 'Participant' },
-    upcomingDiscussions: [
-      {
-        ...mockDiscussions['discussion-1']!,
-        startDateTime: now - 15 * ONE_MINUTE_SECONDS, // Started 15 minutes ago
-        endDateTime: now + 45 * ONE_MINUTE_SECONDS, // Ends in 45 minutes
-        activityDoc: 'https://docs.google.com/document/d/abc123',
-        slackChannelId: 'C1234567890',
-      },
-      mockDiscussions['discussion-2']!,
-    ],
-  },
-  parameters: {
-    docs: {
-      description: {
-        story: 'When a discussion is currently live, the time widget shows "NOW / LIVE" and the primary CTA is "Join now".',
-      },
-    },
-  },
-};
