--- apps/website/src/components/settings/CourseList.test.tsx
@@ -1,621 +1,0 @@
-import '@testing-library/jest-dom';
-import { act, render, screen } from '@testing-library/react';
-import userEvent from '@testing-library/user-event';
-import {
-  afterEach,
-  beforeEach,
-  describe,
-  expect,
-  it,
-  vi,
-} from 'vitest';
-import { createMockCourse, createMockCourseRegistration, createMockMeetPerson } from '../../__tests__/testUtils';
-import CourseList from './CourseList';
-import { TrpcProvider } from '../../__tests__/trpcProvider';
-import { server, trpcMsw } from '../../__tests__/trpcMswSetup';
-
-type MockCourseDetailsProps = {
-  course: { title: string };
-};
-
-type MockDropoutModalProps = {
-  applicantId: string;
-  currentRoundId: string | null;
-  handleClose: () => void;
-};
-
-// Mock the CourseDetails component to avoid testing it here
-vi.mock('./CourseDetails', () => ({
-  default: ({ course }: MockCourseDetailsProps) => (
-    <div aria-label={`Expanded details for ${course.title}`}>Course Details Content</div>
-  ),
-}));
-
-vi.mock('../courses/DropoutModal', () => ({
-  default: ({ applicantId, currentRoundId, handleClose }: MockDropoutModalProps) => (
-    <div role="dialog" aria-label="Dropout or deferral modal">
-      {`Dropout modal for ${applicantId} in ${currentRoundId ?? 'no-round'}`}
-      <button type="button" onClick={handleClose}>Close modal</button>
-    </div>
-  ),
-}));
-
-describe('CourseList', () => {
-  const mockCourse = createMockCourse({
-    id: 'course-1',
-    title: 'Introduction to AI Safety',
-    slug: 'ai-safety',
-  });
-
-  const mockCourseRegistration = createMockCourseRegistration({
-    courseId: 'course-1',
-  });
-
-  const mockMeetPerson = createMockMeetPerson();
-
-  const mockDiscussions = {
-    discussions: [],
-  };
-
-  beforeEach(() => {
-    vi.stubEnv('TZ', 'UTC');
-
-    server.use(
-      trpcMsw.meetPerson.getByCourseRegistrationId.query(() => mockMeetPerson),
-      trpcMsw.groupDiscussions.getByDiscussionIds.query(() => mockDiscussions),
-    );
-  });
-
-  afterEach(() => {
-    vi.unstubAllEnvs();
-  });
-
-  it('renders in-progress course correctly (snapshot)', () => {
-    const inProgressRegistration = {
-      ...mockCourseRegistration,
-      roundStatus: 'Active',
-      certificateCreatedAt: null,
-    };
-
-    const { container } = render(
-      <CourseList courses={[{ course: mockCourse, courseRegistration: inProgressRegistration }]} />,
-      { wrapper: TrpcProvider },
-    );
-    expect(container).toMatchSnapshot();
-  });
-
-  it('renders completed course correctly (snapshot)', () => {
-    const completedRegistration = {
-      ...mockCourseRegistration,
-      roundStatus: 'Past',
-      certificateCreatedAt: new Date('2024-01-01').getTime() / 1000, // Jan 1, 2024 in seconds
-      certificateId: 'cert-123',
-    };
-
-    const { container } = render(
-      <CourseList courses={[{ course: mockCourse, courseRegistration: completedRegistration }]} />,
-      { wrapper: TrpcProvider },
-    );
-    expect(container).toMatchSnapshot();
-  });
-
-  it('starts expanded when startExpanded prop is true', () => {
-    render(
-      <CourseList courses={[{ course: mockCourse, courseRegistration: mockCourseRegistration }]} startExpanded />,
-      { wrapper: TrpcProvider },
-    );
-
-    // Check for collapse button since course is expanded
-    const collapseButtons = screen.getAllByLabelText('Collapse Introduction to AI Safety details');
-    expect(collapseButtons.length).toBeGreaterThan(0);
-    expect(collapseButtons[0]).toHaveAttribute('aria-expanded', 'true');
-
-    // Should show expanded details
-    expect(screen.getByLabelText('Expanded details for Introduction to AI Safety')).toBeInTheDocument();
-  });
-
-  it('starts collapsed when startExpanded prop is false (default)', () => {
-    render(
-      <CourseList courses={[{ course: mockCourse, courseRegistration: mockCourseRegistration }]} />,
-      { wrapper: TrpcProvider },
-    );
-
-    // Check for expand button since course is collapsed
-    const expandButtons = screen.getAllByLabelText('Expand Introduction to AI Safety details');
-    expect(expandButtons.length).toBeGreaterThan(0);
-    expect(expandButtons[0]).toHaveAttribute('aria-expanded', 'false');
-
-    // Should not show expanded details
-    expect(screen.queryByLabelText('Expanded details for Introduction to AI Safety')).not.toBeInTheDocument();
-  });
-
-  it('shows view certificate link for completed course', () => {
-    const completedRegistration = {
-      ...mockCourseRegistration,
-      roundStatus: 'Past',
-      certificateCreatedAt: new Date('2024-01-01').getTime() / 1000, // Jan 1, 2024 in seconds
-      certificateId: 'cert-123',
-    };
-
-    render(
-      <CourseList courses={[{ course: mockCourse, courseRegistration: completedRegistration }]} />,
-      { wrapper: TrpcProvider },
-    );
-
-    const certificateLinks = screen.getAllByRole('link', { name: 'View your certificate' });
-    expect(certificateLinks[0]).toBeInTheDocument();
-    const completedTexts = screen.getAllByText(/Completed on/);
-    expect(completedTexts.length).toBeGreaterThan(0);
-  });
-
-  it('shows dropout or deferral action for active courses even without upcoming discussions', async () => {
-    const user = userEvent.setup();
-    const activeRegistration = {
-      ...mockCourseRegistration,
-      id: 'active-reg',
-      roundId: 'active-round',
-      roundStatus: 'Active',
-      certificateCreatedAt: null,
-    };
-
-    render(
-      <CourseList courses={[{ course: mockCourse, courseRegistration: activeRegistration }]} />,
-      { wrapper: TrpcProvider },
-    );
-
-    const dropoutButtons = await screen.findAllByRole('button', { name: 'Defer or drop out' });
-    expect(dropoutButtons.length).toBeGreaterThan(0);
-
-    await act(async () => {
-      await user.click(dropoutButtons[0]!);
-    });
-
-    expect(screen.getByRole('dialog', { name: 'Dropout or deferral modal' })).toHaveTextContent('Dropout modal for active-reg in active-round');
-  });
-
-  it('shows dropout or deferral action for upcoming non-rejected applications', () => {
-    const futureRegistration = {
-      ...mockCourseRegistration,
-      id: 'future-reg',
-      roundId: 'future-round',
-      roundStatus: 'Future',
-      decision: null,
-      certificateCreatedAt: null,
-    };
-
-    render(
-      <CourseList courses={[{ course: mockCourse, courseRegistration: futureRegistration }]} />,
-      { wrapper: TrpcProvider },
-    );
-
-    expect(screen.getAllByRole('button', { name: 'Defer or drop out' }).length).toBeGreaterThan(0);
-  });
-
-  it('does not show dropout or deferral action for rejected upcoming or past courses', () => {
-    const rejectedFutureRegistration = {
-      ...mockCourseRegistration,
-      id: 'rejected-future-reg',
-      roundStatus: 'Future',
-      decision: 'Reject',
-      certificateCreatedAt: null,
-    };
-    const pastRegistration = {
-      ...mockCourseRegistration,
-      id: 'past-reg',
-      roundStatus: 'Past',
-      certificateCreatedAt: null,
-    };
-
-    render(
-      <CourseList courses={[
-        { course: mockCourse, courseRegistration: rejectedFutureRegistration },
-        { course: createMockCourse({ id: 'course-2', title: 'Past Course', slug: 'past-course' }), courseRegistration: pastRegistration },
-      ]}
-      />,
-      { wrapper: TrpcProvider },
-    );
-
-    expect(screen.queryByRole('button', { name: 'Defer or drop out' })).not.toBeInTheDocument();
-  });
-
-  it('collapses and expands course details', async () => {
-    const user = userEvent.setup();
-    render(
-      <CourseList courses={[{ course: mockCourse, courseRegistration: mockCourseRegistration }]} startExpanded />,
-      { wrapper: TrpcProvider },
-    );
-
-    // Started expanded
-    const collapseButtons = screen.getAllByLabelText('Collapse Introduction to AI Safety details');
-    const collapseButton = collapseButtons[0]!;
-    expect(screen.getByLabelText('Expanded details for Introduction to AI Safety')).toBeInTheDocument();
-
-    // Click to collapse
-    await act(async () => {
-      await user.click(collapseButton);
-    });
-
-    expect(screen.queryByLabelText('Expanded details for Introduction to AI Safety')).not.toBeInTheDocument();
-    expect(collapseButton).toHaveAttribute('aria-expanded', 'false');
-    expect(collapseButton).toHaveAttribute('aria-label', 'Expand Introduction to AI Safety details');
-
-    // Click to expand again
-    await act(async () => {
-      await user.click(collapseButton);
-    });
-
-    expect(screen.getByLabelText('Expanded details for Introduction to AI Safety')).toBeInTheDocument();
-    expect(collapseButton).toHaveAttribute('aria-expanded', 'true');
-    expect(collapseButton).toHaveAttribute('aria-label', 'Collapse Introduction to AI Safety details');
-  });
-
-  it('shows attendance count when course is "Past" with no cert', async () => {
-    const pastNoCertRegistration = {
-      ...mockCourseRegistration,
-      roundStatus: 'Past',
-      certificateCreatedAt: null,
-      certificateId: null,
-    };
-
-    // Mock meetPerson with 5 units but only 2 attended
-    const meetPersonLowAttendance = {
-      ...mockMeetPerson,
-      expectedDiscussionsParticipant: ['disc-1', 'disc-2', 'disc-3', 'disc-4', 'disc-5'],
-      attendedDiscussions: ['disc-1', 'disc-2'],
-      uniqueDiscussionAttendance: 2,
-      numUnits: 5,
-      projectSubmission: null,
-      courseFeedbackForm: 'https://example.com/feedback',
-      courseFeedback: [],
-    };
-
-    const mockExpectedDiscussions = {
-      discussions: [
-        {
-          id: 'disc-1', startDateTime: 1000, endDateTime: 2000, unitNumber: 1,
-        },
-        {
-          id: 'disc-2', startDateTime: 3000, endDateTime: 4000, unitNumber: 2,
-        },
-        {
-          id: 'disc-3', startDateTime: 5000, endDateTime: 6000, unitNumber: 3,
-        },
-        {
-          id: 'disc-4', startDateTime: 7000, endDateTime: 8000, unitNumber: 4,
-        },
-        {
-          id: 'disc-5', startDateTime: 9000, endDateTime: 10000, unitNumber: 5,
-        },
-      ],
-    };
-
-    server.use(
-      trpcMsw.meetPerson.getByCourseRegistrationId.query(() => meetPersonLowAttendance),
-      // @ts-expect-error
-      trpcMsw.groupDiscussions.getByDiscussionIds.query(() => mockExpectedDiscussions),
-    );
-
-    render(
-      <CourseList courses={[{ course: mockCourse, courseRegistration: pastNoCertRegistration }]} />,
-      { wrapper: TrpcProvider },
-    );
-
-    // Wait for the subtitle to appear (multiple elements due to responsive design)
-    const subtitleTexts = await screen.findAllByText(/You attended 2 out of 5 discussions/);
-    expect(subtitleTexts.length).toBeGreaterThan(0);
-
-    // Should show Share feedback button
-    const feedbackButtons = await screen.findAllByRole('link', { name: 'Share feedback' });
-    expect(feedbackButtons.length).toBeGreaterThan(0);
-  });
-
-  it('shows Submit action plan button for any facilitated course without action plan', async () => {
-    const facilitatedCourse = {
-      ...mockCourse,
-      slug: 'agi-strategy',
-    };
-
-    const pastNoCertRegistration = {
-      ...mockCourseRegistration,
-      roundStatus: 'Past',
-      certificateCreatedAt: null,
-      certificateId: null,
-    };
-
-    // Mock meetPerson with no action plan submitted
-    const meetPersonNoActionPlan = {
-      ...mockMeetPerson,
-      expectedDiscussionsParticipant: ['disc-1', 'disc-2', 'disc-3', 'disc-4', 'disc-5'],
-      attendedDiscussions: ['disc-1', 'disc-2', 'disc-3', 'disc-4'],
-      uniqueDiscussionAttendance: 4,
-      numUnits: 5,
-      projectSubmission: null,
-      courseFeedbackForm: 'https://example.com/feedback',
-      courseFeedback: [],
-    };
-
-    const mockExpectedDiscussions = {
-      discussions: [
-        {
-          id: 'disc-1', startDateTime: 1000, endDateTime: 2000, unitNumber: 1,
-        },
-        {
-          id: 'disc-2', startDateTime: 3000, endDateTime: 4000, unitNumber: 2,
-        },
-        {
-          id: 'disc-3', startDateTime: 5000, endDateTime: 6000, unitNumber: 3,
-        },
-        {
-          id: 'disc-4', startDateTime: 7000, endDateTime: 8000, unitNumber: 4,
-        },
-        {
-          id: 'disc-5', startDateTime: 9000, endDateTime: 10000, unitNumber: 5,
-        },
-      ],
-    };
-
-    server.use(
-      trpcMsw.meetPerson.getByCourseRegistrationId.query(() => meetPersonNoActionPlan),
-      // @ts-expect-error
-      trpcMsw.groupDiscussions.getByDiscussionIds.query(() => mockExpectedDiscussions),
-    );
-
-    render(
-      <CourseList courses={[{ course: facilitatedCourse, courseRegistration: pastNoCertRegistration }]} />,
-      { wrapper: TrpcProvider },
-    );
-
-    // Wait for the subtitle showing attendance count
-    const subtitleTexts = await screen.findAllByText(/You attended 4 out of 5 discussions/);
-    expect(subtitleTexts.length).toBeGreaterThan(0);
-
-    // Should show the "Submit action plan" button
-    const actionPlanButtons = await screen.findAllByRole('link', { name: 'Submit action plan' });
-    expect(actionPlanButtons.length).toBeGreaterThan(0);
-    expect(actionPlanButtons[0]).toHaveAttribute('href', expect.stringContaining('miniextensions.com'));
-
-    // Should also show Share feedback button
-    const feedbackButtons = await screen.findAllByRole('link', { name: 'Share feedback' });
-    expect(feedbackButtons.length).toBeGreaterThan(0);
-  });
-
-  it('shows Action plan submitted button when action plan is submitted for any facilitated course', async () => {
-    const facilitatedCourse = {
-      ...mockCourse,
-      slug: 'agi-strategy',
-    };
-
-    const pastNoCertRegistration = {
-      ...mockCourseRegistration,
-      roundStatus: 'Past',
-      certificateCreatedAt: null,
-      certificateId: null,
-    };
-
-    // Mock meetPerson with action plan submitted
-    const meetPersonWithActionPlan = {
-      ...mockMeetPerson,
-      expectedDiscussionsParticipant: ['disc-1', 'disc-2', 'disc-3', 'disc-4', 'disc-5'],
-      attendedDiscussions: ['disc-1', 'disc-2', 'disc-3', 'disc-4'],
-      uniqueDiscussionAttendance: 4,
-      numUnits: 5,
-      projectSubmission: ['action-plan-record-1'],
-      courseFeedbackForm: 'https://example.com/feedback',
-      courseFeedback: [],
-    };
-
-    const mockExpectedDiscussions = {
-      discussions: [
-        {
-          id: 'disc-1', startDateTime: 1000, endDateTime: 2000, unitNumber: 1,
-        },
-        {
-          id: 'disc-2', startDateTime: 3000, endDateTime: 4000, unitNumber: 2,
-        },
-        {
-          id: 'disc-3', startDateTime: 5000, endDateTime: 6000, unitNumber: 3,
-        },
-        {
-          id: 'disc-4', startDateTime: 7000, endDateTime: 8000, unitNumber: 4,
-        },
-        {
-          id: 'disc-5', startDateTime: 9000, endDateTime: 10000, unitNumber: 5,
-        },
-      ],
-    };
-
-    server.use(
-      trpcMsw.meetPerson.getByCourseRegistrationId.query(() => meetPersonWithActionPlan),
-      // @ts-expect-error
-      trpcMsw.groupDiscussions.getByDiscussionIds.query(() => mockExpectedDiscussions),
-    );
-
-    render(
-      <CourseList courses={[{ course: facilitatedCourse, courseRegistration: pastNoCertRegistration }]} />,
-      { wrapper: TrpcProvider },
-    );
-
-    // Wait for the subtitle showing attendance count
-    const subtitleTexts = await screen.findAllByText(/You attended 4 out of 5 discussions/);
-    expect(subtitleTexts.length).toBeGreaterThan(0);
-
-    // Should show disabled "Action plan submitted" button
-    const actionPlanSubmittedButtons = await screen.findAllByRole('button', { name: /Action plan submitted/ });
-    expect(actionPlanSubmittedButtons.length).toBeGreaterThan(0);
-
-    // Should also show Share feedback button
-    const feedbackButtons = await screen.findAllByRole('link', { name: 'Share feedback' });
-    expect(feedbackButtons.length).toBeGreaterThan(0);
-  });
-
-  it('shows locked certificate button when certificate exists but feedback not submitted', async () => {
-    const completedRegistration = {
-      ...mockCourseRegistration,
-      roundStatus: 'Past',
-      certificateCreatedAt: new Date('2024-01-01').getTime() / 1000,
-      certificateId: 'cert-123',
-    };
-
-    // Mock meetPerson with certificate but no feedback submitted
-    const meetPersonNoFeedback = {
-      ...mockMeetPerson,
-      uniqueDiscussionAttendance: 5,
-      numUnits: 5,
-      courseFeedbackForm: 'https://example.com/feedback',
-      courseFeedback: [], // No feedback submitted
-    };
-
-    server.use(trpcMsw.meetPerson.getByCourseRegistrationId.query(() => meetPersonNoFeedback));
-
-    render(
-      <CourseList courses={[{ course: mockCourse, courseRegistration: completedRegistration }]} />,
-      { wrapper: TrpcProvider },
-    );
-
-    // Should show locked certificate button
-    const lockedCertButtons = await screen.findAllByRole('link', { name: /Share feedback to view your certificate/ });
-    expect(lockedCertButtons.length).toBeGreaterThan(0);
-    expect(lockedCertButtons[0]).toHaveAttribute('href', 'https://example.com/feedback');
-  });
-
-  it('shows View your certificate button when certificate exists and feedback submitted', async () => {
-    const completedRegistration = {
-      ...mockCourseRegistration,
-      roundStatus: 'Past',
-      certificateCreatedAt: new Date('2024-01-01').getTime() / 1000,
-      certificateId: 'cert-123',
-    };
-
-    // Mock meetPerson with certificate AND feedback submitted
-    const meetPersonWithFeedback = {
-      ...mockMeetPerson,
-      uniqueDiscussionAttendance: 5,
-      numUnits: 5,
-      courseFeedbackForm: 'https://example.com/feedback',
-      courseFeedback: ['feedback-record-1'], // Feedback submitted
-    };
-
-    server.use(trpcMsw.meetPerson.getByCourseRegistrationId.query(() => meetPersonWithFeedback));
-
-    render(
-      <CourseList courses={[{ course: mockCourse, courseRegistration: completedRegistration }]} />,
-      { wrapper: TrpcProvider },
-    );
-
-    // Should show View your certificate button (not locked)
-    const certificateButtons = await screen.findAllByRole('link', { name: 'View your certificate' });
-    expect(certificateButtons.length).toBeGreaterThan(0);
-    expect(certificateButtons[0]).toHaveAttribute('href', '/certification?id=cert-123');
-  });
-
-  it('shows round name subtitle and hides action plan button for past facilitator course', async () => {
-    const facilitatorRegistration = {
-      ...mockCourseRegistration,
-      role: 'Facilitator' as const,
-      roundStatus: 'Past',
-      roundName: 'AGI Strategy (2025 Aug W35) - Intensive',
-    };
-
-    const meetPersonFacilitator = {
-      ...mockMeetPerson,
-      expectedDiscussionsFacilitator: ['disc-1'],
-      expectedDiscussionsParticipant: [],
-      projectSubmission: null,
-    };
-
-    server.use(trpcMsw.meetPerson.getByCourseRegistrationId.query(() => meetPersonFacilitator));
-
-    render(
-      <CourseList courses={[{ course: { ...mockCourse, slug: 'agi-strategy' }, courseRegistration: facilitatorRegistration }]} />,
-      { wrapper: TrpcProvider },
-    );
-
-    // Should show round name in subtitle
-    const roundNameTexts = await screen.findAllByText(/AGI Strategy \(2025 Aug W35\) - Intensive/);
-    expect(roundNameTexts.length).toBeGreaterThan(0);
-
-    // Should NOT show action plan button for facilitators
-    expect(screen.queryByRole('link', { name: 'Submit action plan' })).not.toBeInTheDocument();
-  });
-
-  it('shows application status text for Future course with pending decision', async () => {
-    const futureRegistration = {
-      ...mockCourseRegistration,
-      roundStatus: 'Future',
-      decision: null,
-      certificateCreatedAt: null,
-      certificateId: null,
-    };
-
-    render(
-      <CourseList courses={[{ course: mockCourse, courseRegistration: futureRegistration }]} roundStartDates={{ 'round-1': '2026-02-09' }} />,
-      { wrapper: TrpcProvider },
-    );
-
-    const statusTexts = await screen.findAllByText('Application in review');
-    expect(statusTexts.length).toBeGreaterThan(0);
-
-    const dateTexts = await screen.findAllByText(/Course starts 9 Feb/);
-    expect(dateTexts.length).toBeGreaterThan(0);
-  });
-
-  it('shows availability button for non-rejected Future course', async () => {
-    const futureRegistration = {
-      ...mockCourseRegistration,
-      roundStatus: 'Future',
-      decision: 'Accept',
-      certificateCreatedAt: null,
-      certificateId: null,
-    };
-
-    render(
-      <CourseList courses={[{ course: mockCourse, courseRegistration: futureRegistration }]} />,
-      { wrapper: TrpcProvider },
-    );
-
-    const availabilityButtons = await screen.findAllByRole('link', { name: 'Submit your availability' });
-    expect(availabilityButtons.length).toBeGreaterThan(0);
-    expect(availabilityButtons[0]).toHaveAttribute('href', expect.stringContaining('availability.bluedot.org'));
-  });
-
-  it('hides availability button for rejected Future course', async () => {
-    const futureRegistration = {
-      ...mockCourseRegistration,
-      roundStatus: 'Future',
-      decision: 'Reject',
-      certificateCreatedAt: null,
-      certificateId: null,
-    };
-
-    render(
-      <CourseList courses={[{ course: mockCourse, courseRegistration: futureRegistration }]} />,
-      { wrapper: TrpcProvider },
-    );
-
-    // Should show rejected status
-    const statusTexts = await screen.findAllByText('Application rejected');
-    expect(statusTexts.length).toBeGreaterThan(0);
-
-    // Should NOT show availability button
-    expect(screen.queryByRole('link', { name: /availability/ })).not.toBeInTheDocument();
-  });
-
-  it('shows View curriculum button for all Future courses', async () => {
-    const futureRegistration = {
-      ...mockCourseRegistration,
-      roundStatus: 'Future',
-      decision: 'Reject',
-      certificateCreatedAt: null,
-      certificateId: null,
-    };
-
-    render(
-      <CourseList courses={[{ course: mockCourse, courseRegistration: futureRegistration }]} />,
-      { wrapper: TrpcProvider },
-    );
-
-    const curriculumButtons = await screen.findAllByRole('link', { name: 'View curriculum' });
-    expect(curriculumButtons.length).toBeGreaterThan(0);
-    expect(curriculumButtons[0]).toHaveAttribute('href', '/courses/ai-safety/1/1');
-  });
-});
