+++ apps/website/src/lib/routes.ts
@@ -105,24 +105,6 @@ const myCourses: BluedotRoute = {
   parentPages: [home],
 };
 
-const legacySettings: BluedotRoute = {
-  title: 'Settings',
-  url: '/legacy/settings',
-  parentPages: [home],
-};
-
-const legacySettingsCourses: BluedotRoute = {
-  title: 'Courses',
-  url: '/legacy/settings/courses',
-  parentPages: [home, legacySettings],
-};
-
-const legacySettingsAccount: BluedotRoute = {
-  title: 'Account',
-  url: '/legacy/settings/account',
-  parentPages: [home, legacySettings],
-};
-
 const subscriptionPreferences: BluedotRoute = {
   title: 'Subscription Preferences',
   url: '/subscription-preferences',
@@ -156,9 +138,6 @@ export const ROUTES = {
   profile,
   account,
   myCourses,
-  legacySettings,
-  legacySettingsCourses,
-  legacySettingsAccount,
   subscriptionPreferences,
 } as const;
