--- apps/website/src/pages/legacy/settings/courses.tsx
@@ -1,135 +1,0 @@
-import type { CourseRegistration } from '@bluedot/db';
-import {
-  CTALinkOrButton, ErrorSection, P, ProgressDots,
-} from '@bluedot/ui';
-import Head from 'next/head';
-import InactiveCourseBanners from '../../../components/courses/InactiveCourseBanners';
-import CourseList from '../../../components/settings/CourseList';
-import SettingsLayout from '../../../components/settings/SettingsLayout';
-import { ROUTES } from '../../../lib/routes';
-import { trpc } from '../../../utils/trpc';
-
-const CURRENT_ROUTE = ROUTES.legacySettingsCourses;
-
-const CoursesSettingsPage = () => {
-  return (
-    <div>
-      <Head>
-        <title>{`${CURRENT_ROUTE.title} | BlueDot Impact`}</title>
-      </Head>
-      <SettingsLayout activeTab="courses" route={CURRENT_ROUTE} afterBreadcrumbs={<InactiveCourseBanners />}>
-        <CoursesContent />
-      </SettingsLayout>
-    </div>
-  );
-};
-
-export default CoursesSettingsPage;
-
-const CoursesContent = () => {
-  const {
-    data: courseRegistrations,
-    isLoading: courseRegistrationsLoading,
-    error: courseRegistrationsError,
-  } = trpc.courseRegistrations.getAll.useQuery();
-
-  const { data: courses, isLoading: coursesLoading, error: coursesError } = trpc.courses.getAll.useQuery();
-
-  const { data: dropoutStatus, isLoading: dropoutsLoading, error: dropoutsError } = trpc.dropout.getStatusForUser.useQuery();
-
-  const isCompleted = (reg: CourseRegistration) => reg.roundStatus !== 'Active' && reg.roundStatus !== 'Future';
-  const isDeferred = (reg: CourseRegistration) => dropoutStatus?.[reg.id]?.isDeferred ?? false;
-  const isDroppedOut = (reg: CourseRegistration) => dropoutStatus?.[reg.id]?.isDroppedOut ?? false;
-
-  // Combine courses and enrollments
-  // eslint-disable-next-line @typescript-eslint/prefer-nullish-coalescing
-  const enrolledCourses = (courseRegistrations || [])
-    .filter((reg) => reg.roundStatus === 'Active' || reg.roundStatus === 'Past' || reg.roundStatus === 'Future' || reg.certificateCreatedAt)
-    // Only show rejected applications while the round is still upcoming (so users can see their status);
-    // hide them once the round is active or past
-    .filter((reg) => reg.decision !== 'Reject' || reg.roundStatus === 'Future')
-    // Exclude dropped out courses (not deferrals)
-    .filter((reg) => !isDroppedOut(reg))
-    .map((courseRegistration) => {
-      const course = courses?.find((c) => c.id === courseRegistration.courseId);
-      return course ? [{ course, courseRegistration }] : [];
-    })
-    .flat();
-
-  // Group courses by status
-  // Upcoming: Future courses (pending, accepted, or rejected applications)
-  const upcomingCourses = enrolledCourses
-    .filter(({ courseRegistration }) => courseRegistration.roundStatus === 'Future');
-
-  // Fetch round start dates for upcoming courses
-  const upcomingRoundIds = [...new Set(upcomingCourses.map(({ courseRegistration }) => courseRegistration.roundId).filter(Boolean))] as string[];
-  const { data: roundStartDates } = trpc.courseRegistrations.getRoundStartDates.useQuery(
-    { roundIds: upcomingRoundIds },
-    { enabled: upcomingRoundIds.length > 0 },
-  );
-
-  // Completed: past courses for participants (non-facilitators), excluding deferred courses
-  const completedCourses = enrolledCourses
-    .filter(({ courseRegistration }) => isCompleted(courseRegistration) && courseRegistration.role !== 'Facilitator' && !isDeferred(courseRegistration))
-    // No-cert courses first (to nudge user to complete), then by completion date descending
-    .sort((a, b) => {
-      const aTime = a.courseRegistration.certificateCreatedAt ?? Infinity;
-      const bTime = b.courseRegistration.certificateCreatedAt ?? Infinity;
-      if (aTime === bTime) return 0;
-      return bTime - aTime;
-    });
-
-  // Facilitated: past courses for facilitators
-  const facilitatedCourses = enrolledCourses
-    .filter(({ courseRegistration }) => isCompleted(courseRegistration) && courseRegistration.role === 'Facilitator');
-
-  // In-progress: Active courses (both participants and facilitators)
-  const inProgressCourses = enrolledCourses.filter(({ courseRegistration }) => courseRegistration.roundStatus === 'Active');
-
-  const loading = courseRegistrationsLoading || coursesLoading || dropoutsLoading;
-  // eslint-disable-next-line @typescript-eslint/prefer-nullish-coalescing
-  const error = courseRegistrationsError || coursesError || dropoutsError;
-
-  if (loading) return <ProgressDots />;
-  if (error) return <ErrorSection error={error} />;
-
-  const totalEnrolledCourses = upcomingCourses.length + inProgressCourses.length + completedCourses.length + facilitatedCourses.length;
-
-  return (
-    <div aria-label="Courses list">
-      <div className="space-y-8">
-        {inProgressCourses.length > 0 && (
-          <section aria-label="In Progress courses" className="lg:mt-2">
-            <P className="font-semibold mb-4">In Progress ({inProgressCourses.length})</P>
-            <CourseList courses={inProgressCourses} startExpanded />
-          </section>
-        )}
-        {upcomingCourses.length > 0 && (
-          <section aria-label="Upcoming courses" className="lg:mt-2">
-            <P className="font-semibold mb-4">Upcoming ({upcomingCourses.length})</P>
-            <CourseList courses={upcomingCourses} roundStartDates={roundStartDates} />
-          </section>
-        )}
-        {completedCourses.length > 0 && (
-          <section aria-label="Completed courses">
-            <P className="font-semibold mb-4">Completed ({completedCourses.length})</P>
-            <CourseList courses={completedCourses} />
-          </section>
-        )}
-        {facilitatedCourses.length > 0 && (
-          <section aria-label="Facilitated courses">
-            <P className="font-semibold mb-4">Facilitated ({facilitatedCourses.length})</P>
-            <CourseList courses={facilitatedCourses} />
-          </section>
-        )}
-      </div>
-
-      {totalEnrolledCourses === 0 && (
-        <div className="flex flex-col gap-4 mt-4">
-          <P>You haven't started any courses yet</P>
-          <CTALinkOrButton url={ROUTES.courses.url}>Join a course</CTALinkOrButton>
-        </div>
-      )}
-    </div>
-  );
-};
